updated uint to ulong for fee and coin

+++ CardanoSharp.Wallet/TransactionBuilding/TransactionBodyBuilder.cs
@@ -8,9 +8,9 @@ namespace CardanoSharp.Wallet.TransactionBuilding
     public interface ITransactionBodyBuilder: IABuilder<TransactionBody>
     {
         ITransactionBodyBuilder AddInput(byte[] transactionId, uint transactionIndex);
-        ITransactionBodyBuilder AddOutput(byte[] address, uint coin, ITokenBundleBuilder tokenBundleBuilder = null);
+        ITransactionBodyBuilder AddOutput(byte[] address, ulong coin, ITokenBundleBuilder tokenBundleBuilder = null);
         ITransactionBodyBuilder SetCertificate(ICertificateBuilder certificateBuilder);
-        ITransactionBodyBuilder SetFee(uint fee);
+        ITransactionBodyBuilder SetFee(ulong fee);
         ITransactionBodyBuilder SetTtl(uint ttl);
         ITransactionBodyBuilder SetMint(ITokenBundleBuilder token);
     }
@@ -43,7 +43,7 @@ namespace CardanoSharp.Wallet.TransactionBuilding
             return this;
         }
 
-        public ITransactionBodyBuilder AddOutput(byte[] address, uint coin, ITokenBundleBuilder tokenBundleBuilder = null)
+        public ITransactionBodyBuilder AddOutput(byte[] address, ulong coin, ITokenBundleBuilder tokenBundleBuilder = null)
         {
             var outputValue = new TransactionOutputValue()
             {
@@ -63,7 +63,7 @@ namespace CardanoSharp.Wallet.TransactionBuilding
             return this;
         }
 
-        public ITransactionBodyBuilder SetFee(uint fee)
+        public ITransactionBodyBuilder SetFee(ulong fee)
         {
             _model.Fee = fee;
             return this;
